Fix user table and seeder

+++ app/Database/Migrations/20210613024900_users.php
@@ -48,7 +48,7 @@ class UserTable extends Migration {
                     'type'           => 'DATE'
             ],
             'contact_number'       => [
-                    'type'           => 'INT',
+                    'type'           => 'DECIMAL',
                     'constraint'     => '11'
             ],
             'email_code'       => [
